Revert "[Prim] add reduce_as op for paddle (#63064)"
This reverts commit 38182e7c4743fe97df999727f31df2f9f71c29e4.
