+++ paddle/phi/api/yaml/backward.yaml
@@ -1900,16 +1900,6 @@
     func : reciprocal_grad
   inplace : (out_grad -> x_grad)
 
-- backward_op : reduce_as_grad
-  forward : reduce_as(Tensor x, Tensor target) -> Tensor(out)
-  args : (Tensor x, Tensor target, Tensor out_grad)
-  output : Tensor(x_grad)
-  infer_meta :
-    func : UnchangedInferMeta
-    param : [x]
-  kernel :
-    func : reduce_as_grad
-
 - backward_op : relu6_grad
   forward : relu6 (Tensor x) -> Tensor(out)
   args : (Tensor out, Tensor out_grad)
