+++ paddle/phi/infermeta/binary.cc
@@ -3055,20 +3055,6 @@ void SequenceMaskInferMeta(const MetaTensor& x,
   y->set_dtype(out_dtype);
 }
 
-void ReduceAsInferMeta(const MetaTensor& x,
-                       const MetaTensor& target,
-                       MetaTensor* out) {
-  DataType out_dtype;
-  if (x.dtype() == DataType::BOOL || x.dtype() == DataType::INT32) {
-    out_dtype = DataType::INT64;
-  } else {
-    out_dtype = x.dtype();
-  }
-  out->set_dtype(out_dtype);
-  out->set_dims(target.dims());
-  out->set_layout(x.layout());
-}
-
 void SoftmaxMaskFuseInferMeta(const MetaTensor& x,
                               const MetaTensor& mask,
                               MetaTensor* out) {
