+++ python/paddle/__init__.py
@@ -475,7 +475,6 @@ from .tensor.math import (  # noqa: F401
     prod,
     rad2deg,
     reciprocal,
-    reduce_as,
     remainder,
     remainder_,
     renorm,
@@ -848,7 +847,6 @@ __all__ = [
     'ones',
     'not_equal',
     'sum',
-    'reduce_as',
     'nansum',
     'nanmean',
     'count_nonzero',
