+++ python/paddle/tensor/__init__.py
@@ -357,7 +357,6 @@ from .math import (  # noqa: F401
     rad2deg,
     reciprocal,
     reciprocal_,
-    reduce_as,
     remainder,
     remainder_,
     renorm,
@@ -526,7 +525,6 @@ tensor_method_func = [
     'square',
     'stanh',
     'sum',
-    'reduce_as',
     'multigammaln',
     'multigammaln_',
     'nan_to_num',
