+++ python/paddle/tensor/math.py
@@ -1578,81 +1578,6 @@ def sum(x, axis=None, dtype=None, keepdim=False, name=None):
         return out
 
 
-def reduce_as(x, target, name=None):
-    """
-    Computes the sum of tensor elements make the shape of its result equal to the shape of target.
-
-    Args:
-        x (Tensor): An N-D Tensor, the data type is bool, float16, float32, float64, int32 or int64.
-        target (Tensor): An N-D Tensor, the length of x shape must greater than or equal to the length of target shape. The data type is bool, float16, float32, float64, int32 or int64.
-
-    Returns:
-        Tensor: The sum of the input tensor x along some axis has the same shape as the shape of the input tensor target, if `x.dtype='bool'`, `x.dtype='int32'`, it's data type is `'int64'`, otherwise it's data type is the same as `x`.
-
-    Examples:
-        .. code-block:: python
-
-            >>> import paddle
-
-            >>> x = paddle.to_tensor([[1, 2, 3, 4], [5, 6, 7, 8]])
-            >>> x
-            Tensor(shape=[2, 4], dtype=int64, place=Place(gpu:0), stop_gradient=True,
-            [[1, 2, 3, 4],
-             [5, 6, 7, 8]])
-            >>> target = paddle.to_tensor([1, 2, 3, 4])
-            >>> target
-            Tensor(shape=[4], dtype=int64, place=Place(gpu:0), stop_gradient=True,
-            [1, 2, 3, 4])
-            >>> res = paddle.reduce_as(x, target)
-            >>> res
-            Tensor(shape=[4], dtype=int64, place=Place(gpu:0), stop_gradient=True,
-            [6 , 8 , 10, 12])
-    """
-
-    if in_dynamic_or_pir_mode():
-        return _C_ops.reduce_as(x, target)
-    else:
-        check_variable_and_dtype(
-            x,
-            'x',
-            [
-                'bool',
-                'uint16',
-                'float16',
-                'float32',
-                'float64',
-                'int16',
-                'int32',
-                'int64',
-            ],
-            'reduce_as',
-        )
-        check_variable_and_dtype(
-            target,
-            'target',
-            [
-                'bool',
-                'uint16',
-                'float16',
-                'float32',
-                'float64',
-                'int16',
-                'int32',
-                'int64',
-            ],
-            'reduce_as',
-        )
-
-        helper = LayerHelper('reduce_as', **locals())
-        out = helper.create_variable_for_type_inference(dtype=x.dtype)
-        helper.append_op(
-            type='reduce_as',
-            inputs={'x': x, 'target': target},
-            outputs={'out': out},
-        )
-        return out
-
-
 def nan_to_num(x, nan=0.0, posinf=None, neginf=None, name=None):
     """
     Replaces NaN, positive infinity, and negative infinity values in input tensor.
