+++ test/legacy_test/test_assign_pos_op.py
@@ -118,6 +118,14 @@ class TestAssignPosAPI(unittest.TestCase):
             )
             assert_allclose(res[0], self.out, self.cum_count)
 
+    def test_api_dygraph(self):
+        paddle.disable_static()
+        x = paddle.to_tensor(self.x)
+        cum_count = paddle.to_tensor(self.cum_count).astype(x.dtype)
+
+        out = utils._assign_pos(x, cum_count)
+        assert_allclose(out.numpy(), self.out, self.cum_count)
+
 
 if __name__ == '__main__':
     paddle.enable_static()
